+++ fpay-api/src/main/java/com/flab/fpay/api/pay/PayController.java
@@ -6,6 +6,8 @@ import com.flab.fpay.api.pay.dto.PaymentDTO;
 import com.flab.fpay.api.pay.dto.PaymentRequestDTO;
 import com.flab.fpay.api.pay.dto.PaymentRequestResDTO;
 import com.flab.fpay.api.pay.dto.PaymentResDTO;
+import com.flab.fpay.api.pay.service.PaymentApproveService;
+import com.flab.fpay.api.pay.service.PaymentCancelService;
 import com.flab.fpay.api.pay.service.PaymentRequestService;
 import com.flab.fpay.api.pay.service.PaymentService;
 import lombok.RequiredArgsConstructor;
@@ -19,6 +21,8 @@ class PayController {
 
     private final PaymentRequestService paymentRequestService;
     private final PaymentService paymentService;
+    private final PaymentApproveService paymentApproveService;
+    private final PaymentCancelService paymentCancelService;
 
     @PostMapping("/ready")
     public ResponseEntity<PaymentRequestResDTO> paymentReady(
@@ -34,8 +38,7 @@ class PayController {
     public ResponseEntity<PaymentResDTO> paymentApprove(
         @RequestBody PaymentDTO paymentDTO) {
 
-        PaymentResDTO paymentResDTO = paymentService.approvePayment(
-            paymentDTO);
+        PaymentResDTO paymentResDTO = paymentApproveService.approvePayment(paymentDTO);
 
         return ResponseEntity.ok(paymentResDTO);
     }
@@ -44,7 +47,7 @@ class PayController {
     public ResponseEntity<PaymentCancelResponseDTO> paymentCancel(
         @RequestBody PaymentCancelRequestDTO paymentCancelRequestDTO) {
 
-        PaymentCancelResponseDTO paymentCancelResponseDTO = paymentService.cancelPayment(
+        PaymentCancelResponseDTO paymentCancelResponseDTO = paymentCancelService.cancelPayment(
             paymentCancelRequestDTO);
 
         return ResponseEntity.ok(paymentCancelResponseDTO);
